+++ lib/src/utils/enum_snake_name_extension.dart
@@ -1,8 +1,5 @@
+import 'package:drift_model_generator/src/utils/make_snake.dart';
+
 extension EnumSnakeName on Enum {
-  String get snakeName => name
-      .replaceAllMapped(
-        RegExp(r'(?<=[a-z])[A-Z]'),
-        (Match m) => '_${m.group(0)}',
-      )
-      .toLowerCase();
+  String get snakeName => makeSnake(name);
 }